backend/requirements.txt
# --- Core framework and async server ---
fastapi==0.116.1
starlette==0.47.1
uvicorn==0.35.0
python-dotenv==1.1.1
PyJWT==2.8.0

# --- Pydantic validation and typing ---
pydantic==2.11.7
pydantic-core==2.33.2
annotated-types==0.7.0
typing-extensions==4.14.1
typing-inspection==0.4.1

# --- Async I/O and runtime support ---
anyio==4.9.0
sniffio==1.3.1
uvloop==0.21.0
httptools==0.6.4
websockets==15.0.1
h11==0.16.0

# --- Linting and formatting (dev only) ---
ruff==0.12.3
black==25.1.0

# --- Testing ---
pytest==8.4.1
iniconfig==2.1.0
pluggy==1.6.0
pygments==2.19.2

# --- CLI and tooling ---
click==8.2.1
watchfiles==1.1.0
pyyaml==6.0.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.3.8
mypy-extensions==1.1.0
idna==3.10

# --- LangChain and AI/ML dependencies ---
langgraph==0.5.4
langchain==0.3.12
langchain-core==0.3.60
langchain-openai==0.2.14
langchain-community==0.3.12
langchain-text-splitters==0.3.3
langchain-chroma==0.2.4
chromadb==1.0.15
openai==1.58.1 
tiktoken==0.7.0

# --- Data processing ---
pandas==2.2.3
numpy==1.26.4

# --- Database and ORM ---
#fastapi==0.103.2
#pydantic==1.10.13
SQLAlchemy==2.0.29
httpx==0.27.0
#uvicorn==0.23.2

